perspective projection

+++ include/triangle.h
@@ -2,7 +2,9 @@
 #define __TRIANGLE_H__
 #include <Eigen/Dense>
 #include "tgaimage.h"
-Eigen::Vector3d barycentric(Eigen::Matrix3i vertices, Eigen::Vector3i point);
-void triangle(Eigen::Matrix3i vertices, TGAImage &image, TGAColor color);
-void triangle(Eigen::Matrix3i vertices, TGAImage &image, TGAColor color, float* zbuffer);
+#include "obj_reader.h"
+Eigen::Vector3d barycentric(Eigen::Matrix3d vertices, Eigen::Vector3i point);
+void triangle(Eigen::Matrix3d vertices, TGAImage &image, TGAColor color);
+void triangle(Eigen::Matrix3d vertices, TGAImage &image, TGAColor color, float* zbuffer);
+// void triangle(Eigen::Matrix3d vertices, TGAImage &image, TGAColor color, float* zbuffer, objReader& obj);
 #endif
